mqtt-screen-standby/requirements.txt
paho-mqtt>=1.3.1